Update upgrading-to-ad-fs-in-windows-server.md
The command to update the connectedserversname parameter in step 5 of the "Upgrade Web Application Proxy" section is incorrect and throws an error when running the command that way. Each WAP server name should be in double quotes. You can also omit the quotes and just go with comma separated strings. but to keep it simple I've made the change with each server name in double quotes.